adds HOC in order to pass cbetData to admin modules
