+++ src/components/admin/layout.js
@@ -1,125 +1,40 @@
-import React from "react"
-import {
-  Container,
-  Row,
-  Col,
-  Nav,
-  Form,
-  FormControl,
-  Navbar,
-  Button,
-} from "react-bootstrap"
-import { Link } from "gatsby"
-import { FaChevronDown } from "react-icons/fa"
+import React, { useState, useEffect } from "react"
+import { Button } from "react-bootstrap"
 import { AzureAD, AuthenticationState } from "react-aad-msal"
 import { signInAuthProvider } from "./../authProvider"
+import AuthenticatedLayout from "./AuthenticatedLayout"
 
-function SideNav() {
-  return (
-    <Col
-      md={2}
-      className="p-0"
-      style={{
-        height: "100vh",
-        color: "whitesmoke",
-        backgroundColor: "#43425d",
-        display: "block !important",
-        overflow: "hidden",
-        position: "sticky",
-        top: "4 rem",
-        zIndex: "1000",
-      }}
-    >
-      <h4 className="pt-2 pb-3 pl-2" style={{ backgroundColor: "#3c3b54" }}>
-        CBET Admin
-      </h4>
-      <Nav defaultActiveKey="/admin" className="flex-column" fill justify>
-        <Link
-          to="/admin"
-          className="nav-link text-light"
-          activeClassName="active"
-        >
-          Dashboard
-        </Link>
-        <Link
-          to="/admin-jobs"
-          className="nav-link text-light"
-          activeClassName="active"
-        >
-          Jobs
-        </Link>
-        <Link
-          to="/admin-events"
-          className="nav-link text-light"
-          activeClassName="active"
-        >
-          Events
-        </Link>
-        <Link
-          to="/admin-blog"
-          className="nav-link text-light"
-          activeClassName="active"
-        >
-          Blog
-        </Link>
-      </Nav>
-    </Col>
-  )
-}
+import useCbetAuth from "../../hooks/use-cbet-auth"
 
 export default function Layout({ title, children, user }) {
+  const authContent = useCbetAuth()
+  const [cbetContent, setCbetContent] = useState([])
+
+  useEffect(() => {
+    // get data from GitHub api
+    fetch(
+      `https://cbetcontent.azurewebsites.net/api/GetCbetContent?code=${authContent}`
+    )
+      .then(response => response.json()) // parse JSON from request
+      .then(resultData => {
+        setCbetContent(resultData)
+      })
+  }, [authContent])
+
   return (
     <AzureAD provider={signInAuthProvider} forceLogin={true}>
       {({ login, logout, authenticationState, error, accountInfo }) => {
         console.log("acct", accountInfo.account.name)
+        console.log("cbet Content is" + cbetContent)
         switch (authenticationState) {
           case AuthenticationState.Authenticated:
             return (
-              <Container fluid>
-                <Row className="flex-xl-nowrap">
-                  <SideNav />
-                  <Col md={10}>
-                    <Navbar expand="lg" variant="light" bg="light">
-                      <Form className="w-75" inline>
-                        <FormControl
-                          type="text"
-                          placeholder="&#x1F50D; Search for jobs, events, and posts"
-                          className="mr-sm-2 w-100"
-                        />
-                      </Form>
-                      <Navbar.Toggle />
-                      <Navbar.Collapse className="justify-content-end">
-                        <Navbar.Text>
-                          Signed in as:
-                          <a href="#login">{` ${accountInfo.account.name}`}</a>
-                        </Navbar.Text>
-                      </Navbar.Collapse>
-                    </Navbar>
-
-                    <Row className="pt-5 pb-5">
-                      <Col md={2}>
-                        <h2>{title}</h2>
-                      </Col>
-                      <Col md={2}>
-                        <Link to="admin-create">
-                          <Button variant="outline-primary">Create New</Button>
-                        </Link>
-                      </Col>
-                      <Col md={2}>
-                        <p>
-                          Filter One <FaChevronDown />
-                        </p>
-                      </Col>
-                      <Col md={2}>
-                        <p>
-                          Filter Two <FaChevronDown />
-                        </p>
-                      </Col>
-                    </Row>
-                    {children}
-                  </Col>
-                </Row>
-              </Container>
+              <AuthenticatedLayout
+                cbetContent={cbetContent}
+                signedInUser={accountInfo.account.name}
+                title={title}
+                children={children}
+              />
             )
           case AuthenticationState.Unauthenticated:
             // TODO: ADD Cbet Styles for unauthenticated
