+++ src/pages/admin.js
@@ -5,7 +5,7 @@ import Layout from "../components/admin/layout"
 const Admin = () => {
   return (
     <Layout title="Dashboard">
-      <SEO title="Admin" />
+      <SEO title="Admin Dashboard" />
       <p>edit the this page from </p>
       <pre className="p-3 mb-2 bg-dark text-white">/src/pages/admin</pre>
     </Layout>
